Retrieve information about car of id 33
in this commit, i retrieve the details of car with the id of 33 from the car dealership inventory. I used this retrieved data to determine the make of the car, the year it was made and also to determine the model of the car. these retrieved information are then console.logged as requested by the car dealer