Nested Object and the this keyword
Create parent object with properties name and age. i made name susan and age 70. also nested a child object with name and age property as well. I set the name of the child object to George and its age to 50. Further creatd a grandchild object with name and age properties and set the name to Sam and age to 30.
The three objects: Parent, child and grandchild objects were then given the speak method which basically outputs a greeting 'hello, my name is [respective name of the object]'