Nested Objects and the this key word II:
This commit is a continutation of the stretch tasks on nested objects and the this keyworrd.
in this commit, the age of the parent object, the name of the child object and the name&age of the grandchild object were all console.logged.

lastly, the speak method was called on all three objects:  the parent, child and grandchild objects.